+++ clients/js/main.ts
@@ -1,19 +1,39 @@
 #!/usr/bin/env node
+
+// <sigh>
+// when the native secp256k1 is missing, the eccrypto library decides TO PRINT A MESSAGE TO STDOUT:
+// https://github.com/bitchan/eccrypto/blob/a4f4a5f85ef5aa1776dfa1b7801cad808264a19c/index.js#L23
+//
+// do you use a CLI tool that depends on that library and try to pipe the output
+// of the tool into another? tough luck
+//
+// for lack of a better way to stop this, we patch the console.info function to
+// drop that particular message...
+// </sigh>
+const info = console.info;
+console.info = function (x) {
+  if (x != "secp256k1 unavailable, reverting to browser version") {
+    info(x);
+  }
+};
+
 import yargs from "yargs";
 
 import { hideBin } from "yargs/helpers";
 
-import { Bech32, fromBech32, toHex } from "@cosmjs/encoding";
+import { fromBech32, toHex } from "@cosmjs/encoding";
+import { CONTRACTS as SDK_CONTRACTS } from "@certusone/wormhole-sdk";
 import {
   isTerraChain,
   assertEVMChain,
-  CONTRACTS,
   setDefaultWasm,
   hexToUint8Array,
   getEmitterAddressSolana,
   getEmitterAddressTerra,
   getEmitterAddressEth,
   getEmitterAddressAlgorand,
+  getEmitterAddressNear,
+  isCosmWasmChain,
 } from "@certusone/wormhole-sdk";
 import { execute_solana } from "./solana";
 import {
@@ -24,7 +44,8 @@ import {
   setStorageAt,
 } from "./evm";
 import { execute_terra } from "./terra";
-import { execute_near } from "./near";
+import { execute_aptos } from "./aptos";
+import { execute_near, upgrade_near, deploy_near } from "./near";
 import * as vaa from "./vaa";
 import { impossible, Payload, serialiseVAA, VAA } from "./vaa";
 import {
@@ -40,12 +61,53 @@ import { ethers } from "ethers";
 import { NETWORKS } from "./networks";
 import base58 from "bs58";
 import { execute_algorand } from "./algorand";
+import { execute_injective } from "./injective";
+import { execute_xpla } from "./xpla";
+import { isOutdated } from "./cmds/update";
 
 setDefaultWasm("node");
+
+if (isOutdated()) {
+  console.error(
+    "\x1b[33m%s\x1b[0m",
+    "WARNING: 'worm' is out of date. Run 'worm update' to update."
+  );
+}
 
 const GOVERNANCE_CHAIN = 1;
 const GOVERNANCE_EMITTER =
   "0000000000000000000000000000000000000000000000000000000000000004";
+
+// TODO: put this into the sdk when things have finalised
+const OVERRIDES = {
+  MAINNET: {
+    aptos: {
+      token_bridge:
+        "0x576410486a2da45eee6c949c995670112ddf2fbeedab20350d506328eefc9d4f",
+      core: "0x5bc11445584a763c1fa7ed39081f1b920954da14e04b32440cba863d03e19625",
+    },
+  },
+  TESTNET: {
+    aptos: {
+      token_bridge:
+        "0x576410486a2da45eee6c949c995670112ddf2fbeedab20350d506328eefc9d4f",
+      core: "0x5bc11445584a763c1fa7ed39081f1b920954da14e04b32440cba863d03e19625",
+    },
+  },
+  DEVNET: {
+    aptos: {
+      token_bridge:
+        "0x84a5f374d29fc77e370014dce4fd6a55b58ad608de8074b0be5571701724da31",
+      core: "0xde0036a9600559e295d5f6802ef6f3f802f510366e0c23912b0655d972166017",
+    },
+  },
+};
+
+export const CONTRACTS = {
+  MAINNET: { ...SDK_CONTRACTS.MAINNET, ...OVERRIDES.MAINNET },
+  TESTNET: { ...SDK_CONTRACTS.TESTNET, ...OVERRIDES.TESTNET },
+  DEVNET: { ...SDK_CONTRACTS.DEVNET, ...OVERRIDES.DEVNET },
+};
 
 function makeVAA(
   emitterChain: number,
@@ -70,6 +132,8 @@ function makeVAA(
 }
 
 yargs(hideBin(process.argv))
+  //TODO(csongor): refactor all commands into the directory structure.
+  .commandDir("cmds")
   ////////////////////////////////////////////////////////////////////////////////
   // Generate
   .command(
@@ -184,6 +248,92 @@ yargs(hideBin(process.argv))
               console.log(serialiseVAA(v));
             }
           )
+          .command(
+            "attestation",
+            "Generate a token attestation VAA",
+            // TODO: putting 'any' here is a workaround for the following error:
+            //
+            //    Type instantiation is excessively deep and possibly infinite.
+            //
+            // The type of the yargs builder grows too big for typescript's
+            // liking, and there's no way to increase the limit. So we
+            // overapproximate with the 'any' type which reduces the typechecking stack.
+            // This is not a great solution, and instead we should move toward
+            // breaking up the commands into multiple modules in the 'cmds' folder.
+            (yargs: any) => {
+              return yargs
+                .option("emitter-chain", {
+                  alias: "e",
+                  describe: "Emitter chain of the VAA",
+                  type: "string",
+                  choices: Object.keys(CHAINS),
+                  required: true,
+                })
+                .option("emitter-address", {
+                  alias: "f",
+                  describe: "Emitter address of the VAA",
+                  type: "string",
+                  required: true,
+                })
+                .option("chain", {
+                  alias: "c",
+                  describe: "Token's chain",
+                  type: "string",
+                  choices: Object.keys(CHAINS),
+                  required: true,
+                })
+                .option("token-address", {
+                  alias: "a",
+                  describe: "Token's address",
+                  type: "string",
+                  required: true,
+                })
+                .option("decimals", {
+                  alias: "d",
+                  describe: "Token's decimals",
+                  type: "number",
+                  required: true,
+                })
+                .option("symbol", {
+                  alias: "s",
+                  describe: "Token's symbol",
+                  type: "string",
+                  required: true,
+                })
+                .option("name", {
+                  alias: "n",
+                  describe: "Token's name",
+                  type: "string",
+                  required: true,
+                });
+            },
+            (argv) => {
+              let emitter_chain = argv["emitter-chain"] as string;
+              assertChain(argv["chain"]);
+              assertChain(emitter_chain);
+              let payload: vaa.TokenBridgeAttestMeta = {
+                module: "TokenBridge",
+                type: "AttestMeta",
+                chain: 0,
+                // TODO: remove these casts (only here because of the workaround above)
+                tokenAddress: parseAddress(
+                  argv["chain"],
+                  argv["token-address"] as string
+                ),
+                tokenChain: toChainId(argv["chain"]),
+                decimals: argv["decimals"] as number,
+                symbol: argv["symbol"] as string,
+                name: argv["name"] as string,
+              };
+              let v = makeVAA(
+                toChainId(emitter_chain),
+                parseAddress(emitter_chain, argv["emitter-address"] as string),
+                argv["guardian-secret"].split(","),
+                payload
+              );
+              console.log(serialiseVAA(v));
+            }
+          )
       );
     },
     (_) => {
@@ -278,9 +428,9 @@ yargs(hideBin(process.argv))
       ) {
         throw Error(`Unknown network: ${network}`);
       }
-      let chain = argv["chain"]
+      let chain = argv["chain"];
       let module = argv["module"] as "Core" | "NFTBridge" | "TokenBridge";
-      let addr = ""
+      let addr = "";
       switch (module) {
         case "Core":
           addr = CONTRACTS[network][chain]["core"];
@@ -295,17 +445,15 @@ yargs(hideBin(process.argv))
           impossible(module);
       }
       if (argv["emitter"]) {
-        if (chain === "solana" || chain === "pythnet") { // TODO: Create an isSolanaChain()
+        if (chain === "solana" || chain === "pythnet") {
+          // TODO: Create an isSolanaChain()
           addr = await getEmitterAddressSolana(addr);
-        } else if (isTerraChain(chain)) {
+        } else if (isCosmWasmChain(chain)) {
           addr = await getEmitterAddressTerra(addr);
         } else if (chain === "algorand") {
           addr = getEmitterAddressAlgorand(BigInt(addr));
         } else if (chain === "near") {
-          if (network !== "MAINNET") {
-            throw Error(`unable to look up near emitter address for ${network}`);
-          }
-          addr = "148410499d3fcda4dcfd68a1ebfcdddda16ab28326448d4aae4d2f0465cdfcb7";
+          addr = await getEmitterAddressNear(addr);
         } else {
           addr = getEmitterAddressEth(addr);
         }
@@ -317,15 +465,14 @@ yargs(hideBin(process.argv))
     "chain-id <chain>",
     "Print the wormhole chain ID integer associated with the specified chain name",
     (yargs) => {
-      return yargs
-        .positional("chain", {
-          describe: "Chain to query",
-          type: "string",
-          choices: Object.keys(CHAINS),
-        });
+      return yargs.positional("chain", {
+        describe: "Chain to query",
+        type: "string",
+        choices: Object.keys(CHAINS),
+      });
     },
     async (argv) => {
-      assertChain(argv["chain"]);   
+      assertChain(argv["chain"]);
       console.log(toChainId(argv["chain"]));
     }
   )
@@ -347,7 +494,6 @@ yargs(hideBin(process.argv))
     },
     async (argv) => {
       assertChain(argv["chain"]);
-      assertEVMChain(argv["chain"]);
       const network = argv.network.toUpperCase();
       if (
         network !== "MAINNET" &&
@@ -359,6 +505,86 @@ yargs(hideBin(process.argv))
       console.log(NETWORKS[network][argv["chain"]].rpc);
     }
   )
+  ////////////////////////////////////////////////////////////////////////////////
+  // Near utilities
+  .command(
+    "near",
+    "NEAR utilites",
+    (yargs) => {
+      return (
+        yargs
+          .option("module", {
+            alias: "m",
+            describe: "Module to query",
+            type: "string",
+            choices: ["Core", "NFTBridge", "TokenBridge"],
+            required: false,
+          })
+          .option("network", {
+            alias: "n",
+            describe: "network",
+            type: "string",
+            choices: ["mainnet", "testnet", "devnet"],
+            required: true,
+          })
+          .option("account", {
+            describe: "near deployment account",
+            type: "string",
+            required: true,
+          })
+          .option("attach", {
+            describe: "attach some near",
+            type: "string",
+            required: false,
+          })
+          .option("target", {
+            describe: "near account to upgrade",
+            type: "string",
+            required: false,
+          })
+          .option("mnemonic", {
+            describe: "near private keys",
+            type: "string",
+            required: false,
+          })
+          .option("keys", {
+            describe: "near private keys",
+            type: "string",
+            required: false,
+          })
+          .command(
+            "contract-update <file>",
+            "Submit a contract update using our specific APIs",
+            (yargs) => {
+              return yargs.positional("file", {
+                type: "string",
+                describe: "wasm",
+              });
+            },
+            async (argv) => {
+              await upgrade_near(argv);
+            }
+          )
+          .command(
+            "deploy <file>",
+            "Submit a contract update using near APIs",
+            (yargs) => {
+              return yargs.positional("file", {
+                type: "string",
+                describe: "wasm",
+              });
+            },
+            async (argv) => {
+              await deploy_near(argv);
+            }
+          )
+      );
+    },
+    (_) => {
+      yargs.showHelp();
+    }
+  )
+
   ////////////////////////////////////////////////////////////////////////////////
   // Evm utilities
   .command(
@@ -554,6 +780,7 @@ yargs(hideBin(process.argv))
     "submit <vaa>",
     "Execute a VAA",
     (yargs) => {
+      // @ts-ignore
       return yargs
         .positional("vaa", {
           describe: "vaa",
@@ -664,13 +891,21 @@ yargs(hideBin(process.argv))
       } else if (chain === "near") {
         await execute_near(parsed_vaa.payload, vaa_hex, network);
       } else if (chain === "injective") {
-        throw Error("INJECTIVE is not supported yet");
+        await execute_injective(parsed_vaa.payload, buf, network);
+      } else if (chain === "xpla") {
+        await execute_xpla(parsed_vaa.payload, buf, network);
       } else if (chain === "osmosis") {
         throw Error("OSMOSIS is not supported yet");
       } else if (chain === "sui") {
         throw Error("SUI is not supported yet");
       } else if (chain === "aptos") {
-        throw Error("APTOS is not supported yet");
+        await execute_aptos(
+          parsed_vaa.payload,
+          buf,
+          network,
+          argv["contract-address"],
+          argv["rpc"]
+        );
       } else if (chain === "wormholechain") {
         throw Error("Wormhole Chain is not supported yet");
       } else {
@@ -679,7 +914,9 @@ yargs(hideBin(process.argv))
         impossible(chain);
       }
     }
-  ).argv;
+  )
+  .strict()
+  .demandCommand().argv;
 
 function hex(x: string): string {
   return ethers.utils.hexlify(x, { allowMissingPrefix: true });
@@ -694,7 +931,7 @@ function parseAddress(chain: ChainName, address: string): string {
     throw Error("Chain unset");
   } else if (isEVMChain(chain)) {
     return "0x" + evm_address(address);
-  } else if (isTerraChain(chain)) {
+  } else if (isCosmWasmChain(chain)) {
     return "0x" + toHex(fromBech32(address).data).padStart(64, "0");
   } else if (chain === "solana" || chain === "pythnet") {
     return "0x" + toHex(base58.decode(address)).padStart(64, "0");
@@ -702,15 +939,14 @@ function parseAddress(chain: ChainName, address: string): string {
     // TODO: is there a better native format for algorand?
     return "0x" + evm_address(address);
   } else if (chain === "near") {
-    return "0x" + hex(address).substring(2).padStart(64, "0")
-  } else if (chain === "injective") {
-    throw Error("INJECTIVE is not supported yet");
+    return "0x" + hex(address).substring(2).padStart(64, "0");
   } else if (chain === "osmosis") {
     throw Error("OSMOSIS is not supported yet");
   } else if (chain === "sui") {
     throw Error("SUI is not supported yet");
   } else if (chain === "aptos") {
-    throw Error("APTOS is not supported yet");
+    // TODO: is there a better native format for aptos?
+    return "0x" + evm_address(address);
   } else if (chain === "wormholechain") {
     return "0x" + tryNativeToHexString(address, chain);
   } else {