+++ clients/js/networks.ts
@@ -82,7 +82,8 @@ const MAINNET = {
     rpc: "https://rpc.mainnet.near.org",
     key: get_env_var("NEAR_KEY"),
     networkId: "mainnet",
-    deployerAccount: "85957f38de1768d6db9eab29bee9dd2a01462aff9c8d83daefb9bcd2506c32d2",
+    deployerAccount:
+      "85957f38de1768d6db9eab29bee9dd2a01462aff9c8d83daefb9bcd2506c32d2",
   },
   injective: {
     rpc: undefined,
@@ -103,6 +104,11 @@ const MAINNET = {
   pythnet: {
     rpc: "http://api.pythnet.pyth.network:8899/",
     key: get_env_var("SOLANA_KEY"),
+  },
+  xpla: {
+    rpc: "https://dimension-lcd.xpla.dev",
+    chain_id: "dimension_37-1",
+    key: get_env_var("XPLA_KEY"),
   },
   wormholechain: {
     rpc: undefined,
@@ -205,21 +211,21 @@ const TESTNET = {
     rpc: "https://rpc.testnet.near.org",
     key: get_env_var("NEAR_KEY_TESTNET"),
     networkId: "testnet",
-    deployerAccount: undefined,
+    deployerAccount: "wormhole.testnet",
   },
   injective: {
     rpc: "https://k8s.testnet.tm.injective.network:443",
     chain_id: "injective-888",
-    key: get_env_var("ETH_KEY_TESTNET"),
+    key: get_env_var("INJECTIVE_KEY_TESTNET"),
   },
   osmosis: {
     rpc: undefined,
     chain_id: "osmo-test-4",
-    key: get_env_var("ETH_KEY_TESTNET"),
+    key: get_env_var("OSMOSIS_KEY_TESTNET"),
   },
   aptos: {
-    rpc: undefined,
-    key: undefined,
+    rpc: "https://fullnode.testnet.aptoslabs.com/v1",
+    key: get_env_var("APTOS_TESTNET"),
   },
   sui: {
     rpc: undefined,
@@ -228,6 +234,11 @@ const TESTNET = {
   pythnet: {
     rpc: "https://api.pythtest.pyth.network/",
     key: get_env_var("SOLANA_KEY_TESTNET"),
+  },
+  xpla: {
+    rpc: "https://cube-lcd.xpla.dev:443",
+    chain_id: "cube_47-5",
+    key: get_env_var("XPLA_KEY_TESTNET"),
   },
   wormholechain: {
     rpc: undefined,
@@ -344,14 +355,19 @@ const DEVNET = {
     rpc: undefined,
     key: undefined,
   },
+  xpla: {
+    rpc: undefined,
+    chain_id: undefined,
+    key: undefined,
+  },
   wormholechain: {
     rpc: "http://localhost:1319",
     chain_id: "wormholechain",
     key: undefined,
   },
   aptos: {
-    rpc: undefined,
-    key: undefined,
+    rpc: "http://0.0.0.0:8080",
+    key: "537c1f91e56891445b491068f519b705f8c0f1a1e66111816dd5d4aa85b8113d",
   },
   sui: {
     rpc: undefined,