+++ sdk/vaa/structs.go
@@ -120,6 +120,10 @@ func (c ChainID) String() string {
 		return "algorand"
 	case ChainIDNear:
 		return "near"
+	case ChainIDAptos:
+		return "aptos"
+	case ChainIDSui:
+		return "sui"
 	case ChainIDEthereumRopsten:
 		return "ethereum-ropsten"
 	case ChainIDKarura:
@@ -144,6 +148,8 @@ func (c ChainID) String() string {
 		return "pythnet"
 	case ChainIDWormchain:
 		return "wormholechain"
+	case ChainIDXpla:
+		return "xpla"
 	default:
 		return fmt.Sprintf("unknown chain ID: %d", c)
 	}
@@ -175,6 +181,10 @@ func ChainIDFromString(s string) (ChainID, error) {
 		return ChainIDAlgorand, nil
 	case "near":
 		return ChainIDNear, nil
+	case "sui":
+		return ChainIDSui, nil
+	case "aptos":
+		return ChainIDAptos, nil
 	case "ethereum-ropsten":
 		return ChainIDEthereumRopsten, nil
 	case "karura":
@@ -199,6 +209,8 @@ func ChainIDFromString(s string) (ChainID, error) {
 		return ChainIDPythNet, nil
 	case "wormholechain":
 		return ChainIDWormchain, nil
+	case "xpla":
+		return ChainIDXpla, nil
 	default:
 		return ChainIDUnset, fmt.Errorf("unknown chain ID: %s", s)
 	}
@@ -244,10 +256,16 @@ const (
 	ChainIDTerra2 ChainID = 18
 	// ChainIDInjective is the ChainID of Injective
 	ChainIDInjective ChainID = 19
+	// ChainIDSui is the ChainID of Sui
+	ChainIDSui ChainID = 21
+	// ChainIDAptos is the ChainID of Aptos
+	ChainIDAptos ChainID = 22
 	// ChainIDArbitrum is the ChainID of Arbitrum
 	ChainIDArbitrum ChainID = 23
 	// ChainIDPythNet is the ChainID of PythNet
 	ChainIDPythNet ChainID = 26
+	// ChainIDXpla is the ChainID of Xpla
+	ChainIDXpla ChainID = 28
 	//ChainIDWormchain is the ChainID of Wormholechain
 	ChainIDWormchain ChainID = 3104