+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.7.2
+### Added
+
+XPLA mainnet support and functions
+
+## 0.7.1
+
+### Added
+
+Neon and XPLA testnet addresses
+
 ## 0.7.0
 
 ### Added